+++ AppCenter/AppCenterTests/MSCustomPropertiesLogTests.m
@@ -6,6 +6,7 @@
 #import "MSCustomPropertiesLog.h"
 #import "MSDevice.h"
 #import "MSTestFrameworks.h"
+#import "MSUtility.h"
 
 @interface MSCustomPropertiesLogTests : XCTestCase
 
@@ -85,8 +86,8 @@
   self.sut.properties = properties;
 
   // When
-  NSData *serializedLog = [NSKeyedArchiver archivedDataWithRootObject:self.sut];
-  id actual = [NSKeyedUnarchiver unarchiveObjectWithData:serializedLog];
+  NSData *serializedLog = [MSUtility archiveKeyedData:self.sut];
+  id actual = [MSUtility unarchiveKeyedData:serializedLog];
 
   // Then
   assertThat(actual, notNilValue());