+++ AppCenterCrashes/AppCenterCrashesTests/MSStackFrameTests.m
@@ -57,8 +57,8 @@
   MSStackFrame *sut = [self stackFrame];
 
   // When
-  NSData *serializedEvent = [NSKeyedArchiver archivedDataWithRootObject:sut];
-  id actual = [NSKeyedUnarchiver unarchiveObjectWithData:serializedEvent];
+  NSData *serializedEvent = [MSUtility archiveKeyedData:sut];
+  id actual = [MSUtility unarchiveKeyedData:serializedEvent];
 
   // Then
   assertThat(actual, notNilValue());