+++ Sasquatch/SasquatchSwift/AppCenterDelegateSwift.swift
@@ -4,8 +4,8 @@
 import AppCenter
 import AppCenterAnalytics
 import AppCenterCrashes
+#if canImport(AppCenterDistribute)
 import AppCenterDistribute
-import AppCenterPush
 
 /**
  * Selectors for reflection.
@@ -17,6 +17,10 @@ import AppCenterPush
   func showDistributeDisabledAlert()
   func delegate() -> MSDistributeDelegate
 }
+#endif
+#if canImport(AppCenterPush)
+import AppCenterPush
+#endif
 
 /**
  * AppCenterDelegate implementation in Swift.
@@ -41,7 +45,11 @@ class AppCenterDelegateSwift: AppCenterDelegate {
   }
 
   func appSecret() -> String {
+#if !targetEnvironment(macCatalyst)
     return kMSSwiftAppSecret
+#else
+    return kMSSwiftCatalystAppSecret
+#endif
   }
 
   func setLogUrl(_ logUrl: String?) {
@@ -78,11 +86,19 @@ class AppCenterDelegateSwift: AppCenterDelegate {
   }
 
   func isDistributeEnabled() -> Bool {
+#if canImport(AppCenterDistribute)
     return MSDistribute.isEnabled()
+#else
+    return false
+#endif
   }
 
   func isPushEnabled() -> Bool {
+#if canImport(AppCenterPush)
     return MSPush.isEnabled()
+#else
+    return false
+#endif
   }
 
   func setAnalyticsEnabled(_ isEnabled: Bool) {
@@ -94,11 +110,15 @@ class AppCenterDelegateSwift: AppCenterDelegate {
   }
 
   func setDistributeEnabled(_ isEnabled: Bool) {
+#if canImport(AppCenterDistribute)
     MSDistribute.setEnabled(isEnabled)
+#endif
   }
 
   func setPushEnabled(_ isEnabled: Bool) {
+#if canImport(AppCenterPush)
     MSPush.setEnabled(isEnabled)
+#endif
   }
 
   // MSAnalytics section.
@@ -153,12 +173,16 @@ class AppCenterDelegateSwift: AppCenterDelegate {
     MSCrashes.generateTestCrash()
   }
 
-  func checkForUpdate() {
-    MSDistribute.checkForUpdate()
-  }
-    
   // MSDistribute section.
+
+  func checkForUpdate() {
+#if canImport(AppCenterDistribute)
+    MSDistribute.checkForUpdate()
+#endif
+  }
+
   func showConfirmationAlert() {
+#if canImport(AppCenterDistribute)
     let sharedInstanceSelector = #selector(Selectors.sharedInstance)
     let confirmationAlertSelector = #selector(Selectors.showConfirmationAlert(_:))
     let releaseDetails = MSReleaseDetails();
@@ -170,9 +194,11 @@ class AppCenterDelegateSwift: AppCenterDelegate {
         _ = distributeInstance.perform(confirmationAlertSelector, with: releaseDetails)
       }
     }
+#endif
   }
 
   func showDistributeDisabledAlert() {
+#if canImport(AppCenterDistribute)
     let sharedInstanceSelector = #selector(Selectors.sharedInstance)
     let disabledAlertSelector = #selector(Selectors.showDistributeDisabledAlert)
     if (MSDistribute.responds(to: sharedInstanceSelector)) {
@@ -181,9 +207,11 @@ class AppCenterDelegateSwift: AppCenterDelegate {
         _ = distributeInstance.perform(disabledAlertSelector)
       }
     }
+#endif
   }
 
   func showCustomConfirmationAlert() {
+#if canImport(AppCenterDistribute)
     let sharedInstanceSelector = #selector(Selectors.sharedInstance)
     let delegateSelector = #selector(Selectors.delegate)
     let releaseDetails = MSReleaseDetails();
@@ -194,6 +222,7 @@ class AppCenterDelegateSwift: AppCenterDelegate {
       let distriuteDelegate = distributeInstance.perform(delegateSelector).takeUnretainedValue()
       _ = distriuteDelegate.distribute?(distributeInstance as? MSDistribute, releaseAvailableWith: releaseDetails)
     }
+#endif
   }
 
   // Last crash report section.