+++ Scripts/build-ios-framework.sh
@@ -4,95 +4,51 @@
 # Licensed under the MIT License.
 set -e
 
-# Load custom build config.
-if [ -r "${SRCROOT}/../.build_config" ]; then
-  source "${SRCROOT}/../.build_config"
-  echo "MS_ARM64E_XCODE_PATH: " $MS_ARM64E_XCODE_PATH
-else
-  echo "Couldn't find custom build config"
-fi
-
 # Sets the target folders and the final framework product.
 TARGET_NAME="${PROJECT_NAME} iOS Framework"
-RESOURCE_BUNDLE="${PROJECT_NAME}Resources"
 
-echo "Building ${TARGET_NAME}."
-
-# Install dir will be the final output to the framework.
-# The following line create it in the root folder of the current project.
+# The directory for final output of the framework.
 PRODUCTS_DIR="${SRCROOT}/../AppCenter-SDK-Apple/iOS"
 
-# The directory to gather all frameworks and build it into xcframework.
-XCFRAMEWORK_DIR="${SRCROOT}/../AppCenter-SDK-Apple/xcframework"
-
-# Working dir will be deleted after the framework creation.
-WORK_DIR=build
-DEVICE_DIR="${WORK_DIR}/Release-iphoneos/"
-SIMULATOR_DIR="${WORK_DIR}/Release-iphonesimulator/"
-
-# Make sure we're inside $SRCROOT.
-cd "${SRCROOT}"
-
-# Cleaning previous build.
-xcodebuild -project "${PROJECT_NAME}.xcodeproj" -configuration "Release" -target "${TARGET_NAME}" clean 
+# Build result paths.
+SCRIPT_BUILD_DIR="${SRCROOT}/build"
+DEVICE_SDK="iphoneos"
+SIMULATOR_SDK="iphonesimulator"
+OUTPUT_DEVICE_DIR="${SCRIPT_BUILD_DIR}/${CONFIGURATION}-${DEVICE_SDK}"
+OUTPUT_SIMULATOR_DIR="${SCRIPT_BUILD_DIR}/${CONFIGURATION}-${SIMULATOR_SDK}"
 
 # Building both architectures.
-xcodebuild -project "${PROJECT_NAME}.xcodeproj" -configuration "Release" -target "${TARGET_NAME}" -sdk iphoneos
-xcodebuild -project "${PROJECT_NAME}.xcodeproj" -configuration "Release" -target "${TARGET_NAME}" -sdk iphonesimulator
+build() {
+  # Print only target name and issues. Mimic Xcode output to make prettify tools happy.
+  echo "=== BUILD TARGET $1 OF PROJECT ${PROJECT_NAME} WITH CONFIGURATION ${CONFIGURATION} ==="
 
-# Cleaning the previous build.
-if [ -d "${PRODUCTS_DIR}/${PROJECT_NAME}.framework" ]; then
-  rm -rf "${PRODUCTS_DIR}/${PROJECT_NAME}.framework"
-fi
+  # Clean building result folder.
+  rm -rf "${SCRIPT_BUILD_DIR}/${CONFIGURATION}-$2"
 
-# Creates and renews the final product folder.
+  # OBJROOT must be customized to avoid conflicts with the current process.
+  xcodebuild OBJROOT="${CONFIGURATION_TEMP_DIR}" PROJECT_TEMP_DIR="${PROJECT_TEMP_DIR}" ONLY_ACTIVE_ARCH=NO \
+    -project "${SRCROOT}/${PROJECT_NAME}.xcodeproj" -configuration "${CONFIGURATION}" -target "$1" -sdk "$2"
+}
+build "${TARGET_NAME}" "${DEVICE_SDK}"
+build "${TARGET_NAME}" "${SIMULATOR_SDK}"
+
+# Clean the previous build and copy the framework.
+rm -rf "${PRODUCTS_DIR}/${PROJECT_NAME}.framework"
 mkdir -p "${PRODUCTS_DIR}"
+cp -RHv "${OUTPUT_DEVICE_DIR}/${PROJECT_NAME}.framework" "${PRODUCTS_DIR}"
 
-# Copy framework.
-cp -R "${DEVICE_DIR}/${PROJECT_NAME}.framework" "${PRODUCTS_DIR}"
+# Uses the Lipo Tool to combine both binary files (i386/x86_64 + armv7/armv7s/arm64/arm64e) into one universal final product.
+echo "Combine binary files into universal final product"
+lipo -create \
+  "${OUTPUT_DEVICE_DIR}/${PROJECT_NAME}.framework/${PROJECT_NAME}" \
+  "${OUTPUT_SIMULATOR_DIR}/${PROJECT_NAME}.framework/${PROJECT_NAME}" \
+  -output "${PRODUCTS_DIR}/${PROJECT_NAME}.framework/${PROJECT_NAME}"
+lipo -info "${PRODUCTS_DIR}/${PROJECT_NAME}.framework/${PROJECT_NAME}"
 
-mkdir -p "${XCFRAMEWORK_DIR}"
-
-# Copy all framework files to use them for xcframework file creation.
-cp -R "${WORK_DIR}/" "${XCFRAMEWORK_DIR}"
-
-# Copy the resource bundle for App Center Distribute.
-if [ -d "${SRCROOT}/${DEVICE_DIR}/${RESOURCE_BUNDLE}.bundle" ]; then
-  echo "Copying resource bundle."
-  cp -R "${SRCROOT}/${DEVICE_DIR}/${RESOURCE_BUNDLE}.bundle" "${PRODUCTS_DIR}" || true
-fi
-
-LIB_IPHONEOS_FINAL="${DEVICE_DIR}/${PROJECT_NAME}.framework/${PROJECT_NAME}"
-
-# Create the arm64e slice in Xcode 10.1 and lipo it with the device binary that was created with oldest supported Xcode version.
-if [ -z "$MS_ARM64E_XCODE_PATH" ] || [ ! -d "$MS_ARM64E_XCODE_PATH" ]; then
-  echo "Environment variable MS_ARM64E_XCODE_PATH not set or not a valid path."
-  echo "Use current Xcode version and lipo -create the fat binary."
-  lipo -create "${DEVICE_DIR}/${PROJECT_NAME}.framework/${PROJECT_NAME}" "${SIMULATOR_DIR}/${PROJECT_NAME}.framework/${PROJECT_NAME}" -output "${PRODUCTS_DIR}/${PROJECT_NAME}.framework/${PROJECT_NAME}"
-else
-
-  # Grep the output of `lipo -archs` if it contains "arm64e". If it does, don't build for arm64e again.
-DOES_CONTAIN_ARM64E=`env DEVELOPER_DIR="$MS_ARM64E_XCODE_PATH" /usr/bin/lipo -archs "${LIB_IPHONEOS_FINAL}" | grep arm64e`
-
-if [ ! -z "${DOES_CONTAIN_ARM64E}" ]; then
-    echo "The binary already contains an arm64e slice."
-else
-    echo "Building the arm64e slice."
-
-    # Move binary that was create with old Xcode to temp location.
-    DEVICE_TEMP_DIR="${DEVICE_DIR}/temp"
-    mkdir -p "${DEVICE_TEMP_DIR}"
-    mv "${DEVICE_DIR}/${PROJECT_NAME}.framework/${PROJECT_NAME}" "${DEVICE_TEMP_DIR}/${PROJECT_NAME}"
-
-    # Build with the Xcode version that supports arm64e.
-    env DEVELOPER_DIR="${MS_ARM64E_XCODE_PATH}" /usr/bin/xcodebuild ARCHS="arm64e" -project "${PROJECT_NAME}.xcodeproj" -configuration "Release" -target "${TARGET_NAME}" 
-
-    # Lipo the binaries that were built with various Xcode versions.
-    env DEVELOPER_DIR="${MS_ARM64E_XCODE_PATH}" lipo -create "${DEVICE_TEMP_DIR}/${PROJECT_NAME}" "${DEVICE_DIR}/${PROJECT_NAME}.framework/${PROJECT_NAME}" -output "${PRODUCTS_DIR}/${PROJECT_NAME}.framework/${PROJECT_NAME}"
-  fi
-
-  echo "Use arm64e Xcode and lipo -create the fat binary."
-  env DEVELOPER_DIR="$MS_ARM64E_XCODE_PATH" lipo -create "${DEVICE_DIR}/${PROJECT_NAME}.framework/${PROJECT_NAME}" "${SIMULATOR_DIR}/${PROJECT_NAME}.framework/${PROJECT_NAME}" -output "${PRODUCTS_DIR}/${PROJECT_NAME}.framework/${PROJECT_NAME}"
-
-# End of arm64e code block.
+# Move the resource bundle outside of framework.
+BUNDLE_NAME="${PROJECT_NAME}Resources.bundle"
+BUNDLE_PATH="${PRODUCTS_DIR}/${PROJECT_NAME}.framework/${BUNDLE_NAME}"
+if [ -e "${BUNDLE_PATH}" ]; then
+  rm -rf "${PRODUCTS_DIR}/${BUNDLE_NAME}"
+  mv -v "${BUNDLE_PATH}" "${PRODUCTS_DIR}"
 fi