+++ Scripts/build-tvos-framework.sh
@@ -7,45 +7,40 @@ set -e
 # Sets the target folders and the final framework product.
 TARGET_NAME="${PROJECT_NAME} tvOS Framework"
 
-echo "Building ${TARGET_NAME}."
+# The directory for final output of the framework.
+PRODUCTS_DIR="${SRCROOT}/../AppCenter-SDK-Apple/tvOS"
 
-# Install dir will be the final output to the framework.
-# The following line create it in the root folder of the current project.
-PRODUCTS_DIR=${SRCROOT}/../AppCenter-SDK-Apple/tvOS
-
-# The directory to gather all frameworks and build it into xcframework.
-XCFRAMEWORK_DIR="${SRCROOT}/../AppCenter-SDK-Apple/xcframework"
-
-# Working dir will be deleted after the framework creation.
-WORK_DIR=build
-DEVICE_DIR="${WORK_DIR}/Release-appletvos/"
-SIMULATOR_DIR="${WORK_DIR}/Release-appletvsimulator/"
-
-# Make sure we're inside $SRCROOT.
-cd "${SRCROOT}"
-
-# Cleaning previous build.
-xcodebuild -project "${PROJECT_NAME}.xcodeproj" -configuration "Release" -target "${TARGET_NAME}" clean
+# Build result paths.
+SCRIPT_BUILD_DIR="${SRCROOT}/build"
+DEVICE_SDK="appletvos"
+SIMULATOR_SDK="appletvsimulator"
+OUTPUT_DEVICE_DIR="${SCRIPT_BUILD_DIR}/${CONFIGURATION}-${DEVICE_SDK}"
+OUTPUT_SIMULATOR_DIR="${SCRIPT_BUILD_DIR}/${CONFIGURATION}-${SIMULATOR_SDK}"
 
 # Building both architectures.
-xcodebuild -project "${PROJECT_NAME}.xcodeproj" -configuration "Release" -target "${TARGET_NAME}" -sdk appletvos 
-xcodebuild -project "${PROJECT_NAME}.xcodeproj" -configuration "Release" -target "${TARGET_NAME}" -sdk appletvsimulator 
+build() {
+  # Print only target name and issues. Mimic Xcode output to make prettify tools happy.
+  echo "=== BUILD TARGET $1 OF PROJECT ${PROJECT_NAME} WITH CONFIGURATION ${CONFIGURATION} ==="
 
-# Cleaning the previous build.
-if [ -d "${PRODUCTS_DIR}/${PROJECT_NAME}.framework" ]; then
-  rm -rf "${PRODUCTS_DIR}/${PROJECT_NAME}.framework"
-fi
+  # Clean building result folder.
+  rm -rf "${SCRIPT_BUILD_DIR}/${CONFIGURATION}-$2"
 
-# Creates and renews the final product folder.
+  # OBJROOT must be customized to avoid conflicts with the current process.
+  xcodebuild OBJROOT="${CONFIGURATION_TEMP_DIR}" PROJECT_TEMP_DIR="${PROJECT_TEMP_DIR}" ONLY_ACTIVE_ARCH=NO \
+    -project "${SRCROOT}/${PROJECT_NAME}.xcodeproj" -configuration "${CONFIGURATION}" -target "$1" -sdk "$2"
+}
+build "${TARGET_NAME}" "${DEVICE_SDK}"
+build "${TARGET_NAME}" "${SIMULATOR_SDK}"
+
+# Clean the previous build and copy the framework.
+rm -rf "${PRODUCTS_DIR}/${PROJECT_NAME}.framework"
 mkdir -p "${PRODUCTS_DIR}"
+cp -RHv "${OUTPUT_DEVICE_DIR}/${PROJECT_NAME}.framework" "${PRODUCTS_DIR}"
 
-# Copy framework.
-cp -R "${DEVICE_DIR}/${PROJECT_NAME}.framework" "${PRODUCTS_DIR}"
-
-mkdir -p "${XCFRAMEWORK_DIR}"
-
-# Copy all framework files to use them for xcframework file creation.
-cp -R "${WORK_DIR}/" "${XCFRAMEWORK_DIR}"
-
-# Uses the Lipo Tool to merge both binary files (i386/x86_64 + armv7/armv7s/arm64) into one Universal final product.
-lipo -create "${DEVICE_DIR}/${PROJECT_NAME}.framework/${PROJECT_NAME}" "${SIMULATOR_DIR}/${PROJECT_NAME}.framework/${PROJECT_NAME}" -output "${PRODUCTS_DIR}/${PROJECT_NAME}.framework/${PROJECT_NAME}"
+# Uses the Lipo Tool to combine both binary files (i386/x86_64 + arm64) into one universal final product.
+echo "Combine binary files into universal final product"
+lipo -create \
+  "${OUTPUT_DEVICE_DIR}/${PROJECT_NAME}.framework/${PROJECT_NAME}" \
+  "${OUTPUT_SIMULATOR_DIR}/${PROJECT_NAME}.framework/${PROJECT_NAME}" \
+  -output "${PRODUCTS_DIR}/${PROJECT_NAME}.framework/${PROJECT_NAME}"
+lipo -info "${PRODUCTS_DIR}/${PROJECT_NAME}.framework/${PROJECT_NAME}"